emui 代码移到lib
