+++ VirtualApp/app/src/main/java/io/virtualapp/VApp.java
@@ -5,7 +5,6 @@ import android.content.Context;
 
 import com.lody.virtual.client.core.VirtualCore;
 
-import io.virtualapp.ui.NotificationHandler;
 import jonathanfinerty.once.Once;
 
 /**
@@ -34,7 +33,6 @@ public class VApp extends Application {
     public void onCreate() {
         gDefault = this;
         super.onCreate();
-        VirtualCore.getCore().setINotificationHandler(new NotificationHandler());
 //        BlockCanary.install(this, new AppBlockCanaryContext()).start();
         VirtualCore.getCore().handleApplication(this);
         if (VirtualCore.getCore().isMainProcess()) {
