+++ VirtualApp/lib/src/main/java/com/lody/virtual/helper/utils/OSUtils.java
@@ -1,0 +1,47 @@
+package com.lody.virtual.helper.utils;
+
+import android.os.Build;
+
+import java.io.IOException;
+import java.lang.reflect.Method;
+
+/**
+ * Created by 247321453 on 2016/7/17.
+ */
+public class OSUtils {
+    private static final String KEY_EMUI_VERSION_CODE = "ro.build.version.emui";
+    private static final String KEY_MIUI_VERSION_CODE = "ro.miui.ui.version.code";
+    private static final String KEY_MIUI_VERSION_NAME = "ro.miui.ui.version.name";
+    private static final String KEY_MIUI_INTERNAL_STORAGE = "ro.miui.internal.storage";
+
+    private static boolean isPropertiesExist(String... keys) {
+        try {
+            BuildProperties prop = BuildProperties.newInstance();
+            for (String key : keys) {
+                String str = prop.getProperty(key);
+                if (str == null)
+                    return false;
+            }
+            return true;
+        } catch (IOException e) {
+            return false;
+        }
+    }
+
+    public static boolean isEMUI() {
+        return isPropertiesExist(KEY_EMUI_VERSION_CODE);
+    }
+
+    public static boolean isMIUI() {
+        return isPropertiesExist(KEY_MIUI_VERSION_CODE, KEY_MIUI_VERSION_NAME, KEY_MIUI_INTERNAL_STORAGE);
+    }
+
+    public static boolean isFlyme() {
+        try {
+            final Method method = Build.class.getMethod("hasSmartBar");
+            return method != null;
+        } catch (final Exception e) {
+            return false;
+        }
+    }
+}
